refactoring(timeline-list): extract styling from container
