+++ web-client/src/modules/timeline/pages/routes/TimelineViewRoute/TimelineViewRoute.tsx
@@ -1,9 +1,19 @@
 import React, { ReactElement } from 'react';
+import styled from 'styled-components';
 
 import TimelineViewContainer from '../../../containers/TimelineViewContainer/TimelineViewContainer';
 
 const TimelineViewRoute: React.FC<any> = ({ match }): ReactElement => (
-  <TimelineViewContainer requestId={match.params.requestId} />
+  <Wrapper>
+    <TimelineViewContainer requestId={match.params.requestId} />
+  </Wrapper>
 );
+
+const Wrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  height: 100%;
+`;
 
 export default TimelineViewRoute;
